backend/requirements.txt
Flask
transformers
torch
faiss-cpu
sentence-transformers
